move DeadlineAlertFields into the new serialization/definitions/deadline
