+++ airflow-core/src/airflow/serialization/definitions/dag.py
@@ -41,7 +41,8 @@ from airflow.models.deadline import Deadline
 from airflow.models.taskinstancekey import TaskInstanceKey
 from airflow.models.tasklog import LogTemplate
 from airflow.observability.stats import Stats
-from airflow.sdk.definitions.deadline import DeadlineReference, DeadlineAlertFields
+from airflow.sdk.definitions.deadline import DeadlineReference
+from airflow.serialization.definitions.deadline import DeadlineAlertFields
 from airflow.serialization.enums import DagAttributeTypes as DAT, Encoding
 from airflow.serialization.definitions.param import SerializedParamsDict
 from airflow.timetables.base import DagRunInfo, DataInterval, TimeRestriction
